Adding packaging for RPMS
Because the RelEng team totally trolled me, I have now added packaging
via RPM to this repository. To utilize said packaing please do the
following:

    make srpm

Then throw that srpm in mock against the target of your choice. In
this case, rpms are kind of dumb, since the golang binary works on
basically any linux that is 64 bit, but whatever. I got trolled.