Add a homebrew formula for installation of vmpool
Since Moses and the rest of RelEng trolled me into packaging with
homebrew, it was time to learn how to use it. It wasn't too bad. The few
caveats were deep copy of git for git describe functionality, and maybe
just general understanding of taps.

Either way, this now works with when installing with HEAD which is what
is recommended. The normal installation also works, but the binary won't
have a version populated, which is kind of drag. I'll have to figure
that out another time.

for usage: clone this project

    brew install --HEAD formula/vmpool.rb