Improve versioning for builds.
By default, this project uses a git describe for versioning. This
doesn't work very well when you have a source tarball (without a .git
directory). So when generating a tarball, a VERSION file will be
written.

If that VERSION file is present, it will be used the VERSION when the
build target is called via make. This should help with Homebrew
building.