Do not require LDAP credentials if a VMPOOL_TOKEN is specified
Previously, due to some odd code-paths and lack of good tests, if you
specified VMPOOL_TOKEN, the CLI utility still needed LDAP_USERNAME and
LDAP_PASSWORD set. This commit fixes that.

There are two caveats.
1. You are unable to list or delete tokens without using LDAP
credentials and that is a design decision in the vmpooler API itself.
2. The API does not allow validation of a token with a token, so we just
assume the token is good.
3. LDAP_USERNAME still needs to be set for vm tagging purposes.

Fixes #7